* fix, bgp.c: free'd pointers set to null in pkt_to_cache_bgp_primitives().   Thanks to Stefano Birmani for his support resolving the issue.

+++ src/plugin_common.c
@@ -501,7 +501,7 @@ void P_eval_historical_acct(struct timeval *stamp, struct timeval *basetime, tim
     if (config.sql_history != COUNT_MONTHLY) {
       int residual;
 
-      residual = timeslot - ((basetime->tv_sec - stamp->tv_sec) % timeslot);
+      residual = timeslot - ((int)(basetime->tv_sec - stamp->tv_sec) % timeslot);
       basetime->tv_sec = stamp->tv_sec - residual;
     }
     else {
